// notebook-angular/src/app/user-profile/account/profile-measurements/models/body-measurements.model.ts
export interface BodyMeasurement {
    id?: number; 
    waistCircumference: number;
    bicepCircumference: number;
    chestCircumference: number;
    thighCircumference: number;
    hipCircumference: number;
    forearmCircumference: number;
    calfCircumference: number;
    neckCircumference: number;
    wristCircumference: number;
    shoulderCircumference: number;
    dateMeasured: string; 
  }
  